@@ -1,6 +1,4 @@
-'use strict';
-
-module.exports = {
+export default {
   semi: true,
   singleQuote: true,
   tabWidth: 2,
