@@ -1,8 +1,6 @@
 import { Token } from '../token';
 import { Chars } from '../chars';
 import { ParserState, Flags } from '../common';
-import { unicodeLookup } from '../unicode';
-import { report, Errors } from '../errors';
 
 export const enum LexerState {
   None = 0,
@@ -33,23 +31,20 @@ export function advanceChar(parser: ParserState): number {
 }
 
 /**
- * Consumes multi unit code point
+ * Consumes possible surrogate pair, return the merged code point or 0
  *
  * @param parser The parser instance
- * @param hi Code point to validate
  */
-export function consumeMultiUnitCodePoint(parser: ParserState, hi: number): 0 | 1 {
+export function consumePossibleSurrogatePair(parser: ParserState): number {
   // See: https://tc39.github.io/ecma262/#sec-ecmascript-language-types-string-type
-  if ((hi & 0xfc00) !== Chars.LeadSurrogateMin) return 0;
+  const hi = parser.currentChar;
+  // Not surrogate
+  if ((hi & 0xfc00) !== Chars.LeadSurrogateMin) return Chars.Null;
   const lo = parser.source.charCodeAt(parser.index + 1);
-  if ((lo & 0xfc00) !== 0xdc00) return 0;
-  hi = parser.currentChar = Chars.NonBMPMin + ((hi & 0x3ff) << 10) + (lo & 0x3ff);
-  if (((unicodeLookup[(hi >>> 5) + 0] >>> hi) & 31 & 1) === 0) {
-    report(parser, Errors.IllegalCharacter, fromCodePoint(hi));
-  }
-  parser.index++;
-  parser.column++;
-  return 1;
+  // Unpaired surrogate
+  if ((lo & 0xfc00) !== Chars.TrailSurrogateMin) return Chars.Null;
+  // Merged code point
+  return Chars.NonBMPMin + ((hi & 0x3ff) << 10) + (lo & 0x3ff);
 }
 
 /**
@@ -86,18 +81,6 @@ export function isExoticECMAScriptWhitespace(ch: number): boolean {
     ch === Chars.ThinSpace ||
     ch === Chars.ByteOrderMark
   );
-}
-
-/**
- * Optimized version of 'fromCodePoint'
- *
- * @param {number} code
- * @returns {string}
- */
-export function fromCodePoint(codePoint: number): string {
-  return codePoint <= 65535
-    ? String.fromCharCode(codePoint)
-    : String.fromCharCode(codePoint >>> 10) + String.fromCharCode(codePoint & 0x3ff);
 }
 
 /**
