@@ -1,10 +1,10 @@
 import { CharFlags, CharTypes } from './charClassifier';
-import { Chars } from '../chars';
 import { Token } from '../token';
 import { ParserState, Context } from '../common';
 import { report, Errors } from '../errors';
-import { advanceChar, LexerState, TokenLookup, scanSingleToken, scanNewLine, consumeLineFeed } from './';
+import { advanceChar, LexerState, scanSingleToken, scanNewLine, consumeLineFeed } from './';
 import { decodeHTMLStrict } from './decodeHTML';
+import { Chars } from '../chars';
 
 /**
  * Scans JSX attribute value
@@ -13,14 +13,16 @@ import { decodeHTMLStrict } from './decodeHTML';
  * @param context Context masks
  */
 export function scanJSXAttributeValue(parser: ParserState, context: Context): Token {
-  parser.startPos = parser.tokenPos = parser.index;
-  parser.startColumn = parser.colPos = parser.column;
-  parser.startLine = parser.linePos = parser.line;
-  parser.token =
+  // skip "=" before the value
+  parser.startIndex = parser.tokenIndex = parser.index;
+  parser.startColumn = parser.tokenColumn = parser.column;
+  parser.startLine = parser.tokenLine = parser.line;
+  parser.setToken(
     CharTypes[parser.currentChar] & CharFlags.StringLiteral
       ? scanJSXString(parser, context)
-      : scanSingleToken(parser, context, LexerState.None);
-  return parser.token;
+      : scanSingleToken(parser, context, LexerState.None)
+  );
+  return parser.getToken();
 }
 
 /**
@@ -41,79 +43,71 @@ export function scanJSXString(parser: ParserState, context: Context): Token {
   if (char !== quote) report(parser, Errors.UnterminatedString);
   parser.tokenValue = parser.source.slice(start, parser.index);
   advanceChar(parser); // skip the quote
-  if (context & Context.OptionsRaw) parser.tokenRaw = parser.source.slice(parser.tokenPos, parser.index);
+  if (context & Context.OptionsRaw) parser.tokenRaw = parser.source.slice(parser.tokenIndex, parser.index);
   return Token.StringLiteral;
 }
 
 /**
- * Scans JSX token
+ * consume Token.LessThan, Token.LeftBrace, or Token.JSXText
  *
  * @param parser The parser object
  */
-export function scanJSXToken(parser: ParserState, context: Context): Token {
-  parser.startPos = parser.tokenPos = parser.index;
-  parser.startColumn = parser.colPos = parser.column;
-  parser.startLine = parser.linePos = parser.line;
+export function nextJSXToken(parser: ParserState, context: Context) {
+  parser.startIndex = parser.tokenIndex = parser.index;
+  parser.startColumn = parser.tokenColumn = parser.column;
+  parser.startLine = parser.tokenLine = parser.line;
 
-  if (parser.index >= parser.end) return (parser.token = Token.EOF);
-
-  const token = TokenLookup[parser.source.charCodeAt(parser.index)];
-
-  switch (token) {
-    // '<'
-    case Token.LessThan: {
-      advanceChar(parser);
-      if (parser.currentChar === Chars.Slash) {
-        advanceChar(parser);
-        parser.token = Token.JSXClose;
-      } else {
-        parser.token = Token.LessThan;
-      }
-
-      break;
-    }
-    // '{'
-    case Token.LeftBrace: {
-      advanceChar(parser);
-      parser.token = Token.LeftBrace;
-      break;
-    }
-    default: {
-      let state = LexerState.None;
-
-      while (parser.index < parser.end) {
-        const type = CharTypes[parser.source.charCodeAt(parser.index)];
-
-        if (type & CharFlags.CarriageReturn) {
-          state |= LexerState.NewLine | LexerState.LastIsCR;
-          scanNewLine(parser);
-        } else if (type & CharFlags.LineFeed) {
-          consumeLineFeed(parser, state);
-          state = (state & ~LexerState.LastIsCR) | LexerState.NewLine;
-        } else {
-          advanceChar(parser);
-        }
-
-        if (CharTypes[parser.currentChar] & CharFlags.JSXToken) break;
-      }
-
-      const raw = parser.source.slice(parser.tokenPos, parser.index);
-      if (context & Context.OptionsRaw) parser.tokenRaw = raw;
-      parser.tokenValue = decodeHTMLStrict(raw);
-      parser.token = Token.JSXText;
-    }
+  if (parser.index >= parser.end) {
+    parser.setToken(Token.EOF);
+    return;
   }
 
-  return parser.token;
+  if (parser.currentChar === Chars.LessThan) {
+    advanceChar(parser);
+    parser.setToken(Token.LessThan);
+    return;
+  }
+
+  if (parser.currentChar === Chars.LeftBrace) {
+    advanceChar(parser);
+    parser.setToken(Token.LeftBrace);
+    return;
+  }
+
+  let state = LexerState.None;
+
+  while (parser.index < parser.end) {
+    const type = CharTypes[parser.source.charCodeAt(parser.index)];
+
+    if (type & CharFlags.CarriageReturn) {
+      state |= LexerState.NewLine | LexerState.LastIsCR;
+      scanNewLine(parser);
+    } else if (type & CharFlags.LineFeed) {
+      consumeLineFeed(parser, state);
+      state = (state & ~LexerState.LastIsCR) | LexerState.NewLine;
+    } else {
+      advanceChar(parser);
+    }
+
+    if (CharTypes[parser.currentChar] & CharFlags.JSXToken) break;
+  }
+
+  // No text, next char is "}" or ">"
+  if (parser.tokenIndex === parser.index) report(parser, Errors.Unexpected);
+
+  const raw = parser.source.slice(parser.tokenIndex, parser.index);
+  if (context & Context.OptionsRaw) parser.tokenRaw = raw;
+  parser.tokenValue = decodeHTMLStrict(raw);
+  parser.setToken(Token.JSXText);
 }
 
 /**
- * Scans JSX identifier
+ * Re-scans JSX identifier which might include hyphen
  *
  * @param parser The parser instance
  */
-export function scanJSXIdentifier(parser: ParserState): Token {
-  if ((parser.token & Token.IsIdentifier) === Token.IsIdentifier) {
+export function rescanJSXIdentifier(parser: ParserState): Token {
+  if ((parser.getToken() & Token.IsIdentifier) === Token.IsIdentifier) {
     const { index } = parser;
     let char = parser.currentChar;
     while (CharTypes[char] & (CharFlags.Hyphen | CharFlags.IdentifierPart)) {
@@ -121,6 +115,6 @@ export function scanJSXIdentifier(parser: ParserState): Token {
     }
     parser.tokenValue += parser.source.slice(index, parser.index);
   }
-  parser.token = Token.Identifier;
-  return parser.token;
+  parser.setToken(Token.Identifier, true);
+  return parser.getToken();
 }
