@@ -25,8 +25,13 @@ describe('Lexer - String', () => {
     [Context.None, Token.StringLiteral, '""', ''],
     [Context.None, Token.StringLiteral, '"123"', '123'],
     [Context.None, Token.StringLiteral, '"true"', 'true'],
-    [Context.None, Token.StringLiteral, '"\
-    "', '    '],
+    [
+      Context.None,
+      Token.StringLiteral,
+      '"\
+    "',
+      '    '
+    ],
 
     // Russian letters
     [Context.None, Token.StringLiteral, '"\\б"', 'б'],
@@ -62,10 +67,10 @@ describe('Lexer - String', () => {
     [Context.None, Token.StringLiteral, '"\\uf2ff"', ''],
     [Context.None, Token.StringLiteral, '"\\u0123"', 'ģ'],
     [Context.None, Token.StringLiteral, '"\\u0123 postfix"', 'ģ postfix'],
-    [Context.None, Token.StringLiteral, '"\\u{89abc}"', 'Ȧʼ'],
+    [Context.None, Token.StringLiteral, '"\\u{89abc}"', '򉪼'],
     [Context.None, Token.StringLiteral, '"\\u{CDEF}"', '췯'],
-    [Context.None, Token.StringLiteral, '"\\u{0000000000000000000010ffff}"', 'пϿ'],
-    [Context.None, Token.StringLiteral, '"\\u{10ffff}"', 'пϿ'],
+    [Context.None, Token.StringLiteral, '"\\u{0000000000000000000010ffff}"', '􏿿'],
+    [Context.None, Token.StringLiteral, '"\\u{10ffff}"', '􏿿'],
     [Context.None, Token.StringLiteral, '"\\u0062"', 'b'],
     [Context.None, Token.StringLiteral, '"\\u0410"', 'А'],
     [Context.None, Token.StringLiteral, '"\\u0412"', 'В'],
