@@ -6,29 +6,27 @@ import { scanSingleToken } from '../../src/lexer/scan';
 
 describe('Lexer - Unicode Escape', () => {
   const tokens: [Context, Token, string, string][] = [
-    [Context.None, Token.Identifier, '\\u0049', 'I'],
-    [Context.None, Token.Identifier, '\\u0627\\u0644\\u0642\\u0637', 'القط'],
-    [Context.None, Token.Identifier, '\\u0066', 'f'],
-    [Context.None, Token.Identifier, '\\u0066', 'f'],
-    [Context.None, Token.Identifier, '\\u0061', 'a'],
-    [Context.None, Token.Identifier, '\\u{61}', 'a'],
-    [Context.None, Token.Identifier, '\\u{0000061}', 'a'],
-    [Context.None, Token.Identifier, '\\u{0066}', 'f'],
-    [Context.None, Token.LetKeyword, 'l\\u0065t', 'let'],
-    [Context.None, Token.Identifier, '\\u{0000000066}', 'f'],
-    [Context.None, Token.Identifier, '\\u{000000000000000066}', 'f'],
-    [Context.None, Token.Identifier, '\\u{00000000000000000000000066}', 'f'],
-    [Context.None, Token.Identifier, '\\u0062\\u0066', 'bf'],
-    [Context.None, Token.Identifier, '\\u0067', 'g'],
-    [Context.None, Token.Identifier, '\\u0062\\u0066', 'bf'],
-    [Context.None, Token.Identifier, '\\u0062\\u0066', 'bf'],
-    [Context.None, Token.Identifier, '\\u01F602', 'Ƕ02'],
-    [Context.None, Token.Identifier, '\\u{41}', 'A'],
-    [Context.None, Token.Identifier, '\\u{0041}', 'A'],
-    [Context.None, Token.Identifier, '\\u{03BB}', 'λ'],
-    [Context.None, Token.Identifier, '\\u{6728}', '木'],
-    [Context.None, Token.Identifier, '\\u{7800}', '砀'],
-    [Context.None, Token.Identifier, '\\u{4b00}', '䬀']
+    [Context.None, Token.Identifier | Token.IsEscaped, '\\u0049', 'I'],
+    [Context.None, Token.Identifier | Token.IsEscaped, '\\u0627\\u0644\\u0642\\u0637', 'القط'],
+    [Context.None, Token.Identifier | Token.IsEscaped, '\\u0066', 'f'],
+    [Context.None, Token.Identifier | Token.IsEscaped, '\\u0066', 'f'],
+    [Context.None, Token.Identifier | Token.IsEscaped, '\\u0061', 'a'],
+    [Context.None, Token.Identifier | Token.IsEscaped, '\\u{61}', 'a'],
+    [Context.None, Token.Identifier | Token.IsEscaped, '\\u{0000061}', 'a'],
+    [Context.None, Token.Identifier | Token.IsEscaped, '\\u{0066}', 'f'],
+    [Context.None, Token.LetKeyword | Token.IsEscaped, 'l\\u0065t', 'let'],
+    [Context.None, Token.Identifier | Token.IsEscaped, '\\u{0000000066}', 'f'],
+    [Context.None, Token.Identifier | Token.IsEscaped, '\\u{000000000000000066}', 'f'],
+    [Context.None, Token.Identifier | Token.IsEscaped, '\\u{00000000000000000000000066}', 'f'],
+    [Context.None, Token.Identifier | Token.IsEscaped, '\\u0062\\u0066', 'bf'],
+    [Context.None, Token.Identifier | Token.IsEscaped, '\\u0067', 'g'],
+    [Context.None, Token.Identifier | Token.IsEscaped, '\\u01F602', 'Ƕ02'],
+    [Context.None, Token.Identifier | Token.IsEscaped, '\\u{41}', 'A'],
+    [Context.None, Token.Identifier | Token.IsEscaped, '\\u{0041}', 'A'],
+    [Context.None, Token.Identifier | Token.IsEscaped, '\\u{03BB}', 'λ'],
+    [Context.None, Token.Identifier | Token.IsEscaped, '\\u{6728}', '木'],
+    [Context.None, Token.Identifier | Token.IsEscaped, '\\u{7800}', '砀'],
+    [Context.None, Token.Identifier | Token.IsEscaped, '\\u{4b00}', '䬀']
   ];
 
   for (const [ctx, token, op, value] of tokens) {
