@@ -152,8 +152,6 @@ describe('Declarations - Function', () => {
     ['o = {foo(x = implements = y){ "use strict"; }}', Context.None],
     ['o = {foo(x= eval = y){ "use strict"; }}', Context.None],
     ['function f(async function() {}) { }', Context.None],
-    ['function foo(p\\u0061ckage) { "use strict"; }', Context.None],
-    ['function foo(p\\u0061ckage) { }', Context.Strict],
     ['function await() {}', Context.Strict | Context.Module],
     ['function *await() {}', Context.Strict | Context.Module],
     ['function foo(package) { "use strict"; }', Context.None],
@@ -206,7 +204,8 @@ describe('Declarations - Function', () => {
     ['function f(){ class x { foo(x=yield y){} }  }', Context.None],
     ['function f(){ class x { foo(x=new (yield)()){} }  }', Context.None],
     ['function f(){ class x { [yield](){} }  }', Context.None],
-    ['function f(){ class x { [yield y](){} }  }', Context.None]
+    ['function f(){ class x { [yield y](){} }  }', Context.None],
+    ['function foo(p\\u0061ckage) { "use strict"; }', Context.None]
   ]);
 
   for (const arg of [
@@ -563,7 +562,8 @@ describe('Declarations - Function', () => {
     'function *f(){ class x { [yield](){} }  }',
     'function *f(){ class x { yield(){} }  }',
     'function f() { throw `${delete(y)}`; }',
-    'async function* a() { for (let m in ((yield))) x;  (r = a) => {} }'
+    'async function* a() { for (let m in ((yield))) x;  (r = a) => {} }',
+    'function foo(p\\u0061ckage) { }'
   ]) {
     it(`${arg}`, () => {
       t.doesNotThrow(() => {
@@ -761,7 +761,8 @@ describe('Declarations - Function', () => {
             expression: {
               type: 'Literal',
               value: 'use strict'
-            }
+            },
+            directive: 'use strict'
           },
           {
             type: 'FunctionDeclaration',
@@ -1926,6 +1927,7 @@ describe('Declarations - Function', () => {
                       type: 'Literal',
                       value: 'use strict'
                     },
+                    directive: 'use strict',
                     type: 'ExpressionStatement'
                   }
                 ],
