@@ -65,7 +65,13 @@ describe('Expressions - Object', () => {
     'with (Math) { ({x=1}) };',
     'true ? {x=1} : 1;',
     'false ? 1 : {x=1};',
-    '{x=1} ? 2 : 3;'
+    '{x=1} ? 2 : 3;',
+    '({static a() {}})',
+    '({static b})',
+    '({a b})',
+    '({a b() {}})',
+    '{get async(v) {}}',
+    '{get let(v) {}}'
   ]) {
     it(`${arg}`, () => {
       t.throws(() => {
@@ -173,9 +179,15 @@ describe('Expressions - Object', () => {
     '{d: {}[x ? y : z] += a}',
     '{ b: c.d === e ? f : g }',
     '{ "b": c.d === e ? f : g }',
-    '{ [b]: c.d === e ? f : g }'
+    '{ [b]: c.d === e ? f : g }',
+    '{async static() {}}',
+    '{async *static() {}}',
+    '{set async(v) {}}',
+    '{get async() {}}',
+    '{get let() {}}',
+    'async ({__proto__: a, __proto__: b}) => 1'
   ]) {
-    it(`'use strict'; x = ${arg}`, () => {
+    it(`x = ${arg}`, () => {
       t.doesNotThrow(() => {
         parseSource(`x = ${arg}`, undefined, Context.OptionsWebCompat);
       });
@@ -210,7 +222,8 @@ describe('Expressions - Object', () => {
     'x = {__proto__: 1, "__proto__": 2}',
     'x = {\'__proto__\': 1, "__proto__": 2}',
     "x = {'__proto__': 1, __proto__: 2}",
-    'x = {__proto__: 1, "__proto__": 2}'
+    'x = {__proto__: 1, "__proto__": 2}',
+    'async ({__proto__: a, __proto__: b});'
   ]) {
     it(`${arg}`, () => {
       t.throws(() => {
@@ -219,7 +232,7 @@ describe('Expressions - Object', () => {
     });
 
     it(`${arg}`, () => {
-      t.doesNotThrow(() => {
+      t.throws(() => {
         parseSource(`${arg}`, undefined, Context.None);
       });
     });
@@ -232,25 +245,25 @@ describe('Expressions - Object', () => {
     '__proto__: {}, set __proto__(v) {}',
     '__proto__: {}, __proto__'
   ]) {
-    it(`${arg}`, () => {
+    it(`({${arg}})`, () => {
       t.doesNotThrow(() => {
         parseSource(`({${arg}});`, undefined, Context.None);
       });
     });
 
-    it(`"use strict"; (${arg})`, () => {
+    it(`"use strict"; ({${arg}});`, () => {
       t.doesNotThrow(() => {
         parseSource(`"use strict"; ({${arg}});`, undefined, Context.None);
       });
     });
 
-    it(`x = ${arg}`, () => {
+    it(`x = {${arg}}`, () => {
       t.doesNotThrow(() => {
         parseSource(`x = {${arg}};`, undefined, Context.None);
       });
     });
 
-    it(`x = ${arg}`, () => {
+    it(`x = {${arg}}`, () => {
       t.doesNotThrow(() => {
         parseSource(`x = {${arg}};`, undefined, Context.OptionsWebCompat);
       });
@@ -922,7 +935,6 @@ describe('Expressions - Object', () => {
     'key: bar.foo + x',
     'key: bar.foo + x',
     'async: async.await + x',
-    'key: await /foo/g',
     'key: bar/x',
     'key: bar, foo: zoo',
     'x:y} = { ',
@@ -1098,7 +1110,6 @@ describe('Expressions - Object', () => {
     ['({...(a,b)}) => {}', Context.None],
     ['x = {\'__proto__\': 1, "__proto__": 2}', Context.OptionsWebCompat],
     ['x = {__proto__: 1, "__proto__": 2}', Context.OptionsWebCompat],
-    ["x = {'__proto__': 1, __proto__: 2}", Context.OptionsWebCompat],
     ['({ get *x(){} })', Context.None],
     ['({get +:3})', Context.None],
     ['({get bar(x) {})', Context.None],
@@ -1406,7 +1417,10 @@ describe('Expressions - Object', () => {
     ['let {...x, ...y} = {}', Context.None],
     ['({...rest, b} = {})', Context.None],
     ["x = {'__proto__': 1, __proto__: 2}", Context.OptionsWebCompat],
-    ['({g\\u0065t m() {} });', Context.None]
+    ["x = {'__proto__': 1, __proto__: 2}", Context.None],
+    ['({g\\u0065t m() {} });', Context.None],
+    ['([{web: false, __proto__: a, __proto__: b}]);', Context.None],
+    ['({web: false, __proto__: a, __proto__: b});', Context.None]
   ]);
 
   pass('Expressions - Object (pass)', [
@@ -1483,193 +1497,6 @@ describe('Expressions - Object', () => {
         start: 0,
         end: 20,
         range: [0, 20]
-      }
-    ],
-    [
-      '({web: false, __proto__: a, __proto__: b});',
-      Context.None,
-      {
-        body: [
-          {
-            expression: {
-              properties: [
-                {
-                  computed: false,
-                  key: {
-                    name: 'web',
-                    type: 'Identifier'
-                  },
-                  kind: 'init',
-                  method: false,
-                  shorthand: false,
-                  type: 'Property',
-                  value: {
-                    type: 'Literal',
-                    value: false
-                  }
-                },
-                {
-                  computed: false,
-                  key: {
-                    name: '__proto__',
-                    type: 'Identifier'
-                  },
-                  kind: 'init',
-                  method: false,
-                  shorthand: false,
-                  type: 'Property',
-                  value: {
-                    name: 'a',
-                    type: 'Identifier'
-                  }
-                },
-                {
-                  computed: false,
-                  key: {
-                    name: '__proto__',
-                    type: 'Identifier'
-                  },
-                  kind: 'init',
-                  method: false,
-                  shorthand: false,
-                  type: 'Property',
-                  value: {
-                    name: 'b',
-                    type: 'Identifier'
-                  }
-                }
-              ],
-              type: 'ObjectExpression'
-            },
-            type: 'ExpressionStatement'
-          }
-        ],
-        sourceType: 'script',
-        type: 'Program'
-      }
-    ],
-    [
-      '([{web: false, __proto__: a, __proto__: b}]);',
-      Context.None,
-      {
-        body: [
-          {
-            expression: {
-              elements: [
-                {
-                  properties: [
-                    {
-                      computed: false,
-                      key: {
-                        name: 'web',
-                        type: 'Identifier'
-                      },
-                      kind: 'init',
-                      method: false,
-                      shorthand: false,
-                      type: 'Property',
-                      value: {
-                        type: 'Literal',
-                        value: false
-                      }
-                    },
-                    {
-                      computed: false,
-                      key: {
-                        name: '__proto__',
-                        type: 'Identifier'
-                      },
-                      kind: 'init',
-                      method: false,
-                      shorthand: false,
-                      type: 'Property',
-                      value: {
-                        name: 'a',
-                        type: 'Identifier'
-                      }
-                    },
-                    {
-                      computed: false,
-                      key: {
-                        name: '__proto__',
-                        type: 'Identifier'
-                      },
-                      kind: 'init',
-                      method: false,
-                      shorthand: false,
-                      type: 'Property',
-                      value: {
-                        name: 'b',
-                        type: 'Identifier'
-                      }
-                    }
-                  ],
-                  type: 'ObjectExpression'
-                }
-              ],
-              type: 'ArrayExpression'
-            },
-            type: 'ExpressionStatement'
-          }
-        ],
-        sourceType: 'script',
-        type: 'Program'
-      }
-    ],
-    [
-      "x = {'__proto__': 1, __proto__: 2}",
-      Context.None,
-      {
-        body: [
-          {
-            expression: {
-              left: {
-                name: 'x',
-                type: 'Identifier'
-              },
-              operator: '=',
-              right: {
-                properties: [
-                  {
-                    computed: false,
-                    key: {
-                      type: 'Literal',
-                      value: '__proto__'
-                    },
-                    kind: 'init',
-                    method: false,
-                    shorthand: false,
-                    type: 'Property',
-                    value: {
-                      type: 'Literal',
-                      value: 1
-                    }
-                  },
-                  {
-                    computed: false,
-                    key: {
-                      name: '__proto__',
-                      type: 'Identifier'
-                    },
-                    kind: 'init',
-                    method: false,
-                    shorthand: false,
-                    type: 'Property',
-                    value: {
-                      type: 'Literal',
-                      value: 2
-                    }
-                  }
-                ],
-                type: 'ObjectExpression'
-              },
-              type: 'AssignmentExpression'
-            },
-            type: 'ExpressionStatement'
-          }
-        ],
-        sourceType: 'script',
-        type: 'Program'
       }
     ],
     [
@@ -2213,69 +2040,6 @@ describe('Expressions - Object', () => {
           }
         ],
         sourceType: 'script'
-      }
-    ],
-    [
-      '({ __proto__: null, other: null, "__proto__": null });',
-      Context.None,
-      {
-        body: [
-          {
-            expression: {
-              properties: [
-                {
-                  computed: false,
-                  key: {
-                    name: '__proto__',
-                    type: 'Identifier'
-                  },
-                  kind: 'init',
-                  method: false,
-                  shorthand: false,
-                  type: 'Property',
-                  value: {
-                    type: 'Literal',
-                    value: null
-                  }
-                },
-                {
-                  computed: false,
-                  key: {
-                    name: 'other',
-                    type: 'Identifier'
-                  },
-                  kind: 'init',
-                  method: false,
-                  shorthand: false,
-                  type: 'Property',
-                  value: {
-                    type: 'Literal',
-                    value: null
-                  }
-                },
-                {
-                  computed: false,
-                  key: {
-                    type: 'Literal',
-                    value: '__proto__'
-                  },
-                  kind: 'init',
-                  method: false,
-                  shorthand: false,
-                  type: 'Property',
-                  value: {
-                    type: 'Literal',
-                    value: null
-                  }
-                }
-              ],
-              type: 'ObjectExpression'
-            },
-            type: 'ExpressionStatement'
-          }
-        ],
-        sourceType: 'script',
-        type: 'Program'
       }
     ],
     [
