@@ -20,7 +20,9 @@ describe('Lexical - Block', () => {
     ['{ async function f() {} async function f() {} }', Context.OptionsLexical],
     ['{ async function f() {} class f {} }', Context.OptionsLexical],
     ['{ async function f() {} function f() {} }', Context.OptionsLexical],
+    ['{ async function f() {} function f() {} }', Context.OptionsLexical | Context.OptionsWebCompat],
     ['{ async function f() {} function* f() {} }', Context.OptionsLexical],
+    ['{ async function f() {} function* f() {} }', Context.OptionsLexical | Context.OptionsWebCompat],
     ['{ const f = 0; function* f() {} }', Context.OptionsLexical],
     ['function x() { { async function* f() {}; var f; } }', Context.OptionsLexical],
     ['function x() { { const f = 0; var f; } }', Context.OptionsLexical],
@@ -39,6 +41,7 @@ describe('Lexical - Block', () => {
     ['{ async function f() {} var f }', Context.OptionsLexical],
     ['{ function a(){} function a(){} }', Context.OptionsLexical],
     ['{ function f() {} async function* f() {} }', Context.OptionsLexical],
+    ['{ function f() {} async function* f() {} }', Context.OptionsLexical | Context.OptionsWebCompat],
     ['{ async function* f() {} async function* f() {} }', Context.OptionsLexical],
     ['{ async function* f() {} const f = 0 }', Context.OptionsLexical],
     ['{ { var f; } async function f() {}; }', Context.OptionsLexical],
@@ -77,6 +80,7 @@ describe('Lexical - Block', () => {
     ['for (const x in {}) { var x; }', Context.OptionsLexical],
     ['{ async function f() {} let f }', Context.OptionsLexical],
     ['{ async function* f() {} async function f() {} }', Context.OptionsLexical],
+    ['{ async function* f() {} async function f() {} }', Context.OptionsLexical | Context.OptionsWebCompat],
     ['{ async function* f() {} var f }', Context.OptionsLexical],
     ['{ class f {} const f = 0 }', Context.OptionsLexical],
     ['{ const f = 0; async function f() {} }', Context.OptionsLexical],
@@ -84,6 +88,7 @@ describe('Lexical - Block', () => {
     ['{ function* f() {} class f {} }', Context.OptionsLexical],
     ['{ for (var x;;); const x = 1 }', Context.OptionsLexical],
     ['{ function* f() {} async function f() {} }', Context.OptionsLexical],
+    ['{ function* f() {} async function f() {} }', Context.OptionsLexical | Context.OptionsWebCompat],
     ['function g() { { function f() {} { var f; } }}', Context.OptionsLexical],
     ['function x() { { function* f() {}; var f; } }', Context.OptionsLexical],
     ['function x() { { async function* f() {}; var f; } }', Context.OptionsLexical],
@@ -93,8 +98,10 @@ describe('Lexical - Block', () => {
     ['{ function foo() {} function foo() {} }', Context.OptionsLexical],
     ['{ let a; { var a; } }', Context.OptionsLexical],
     ['{ async function f() {} async function* f() {} }', Context.OptionsLexical],
+    ['{ async function f() {} async function* f() {} }', Context.OptionsLexical | Context.OptionsWebCompat],
     ['{ async function f() {} const f = 0 }', Context.OptionsLexical],
     ['{ async function* f() {} function f() {} }', Context.OptionsLexical],
+    ['{ async function* f() {} function f() {} }', Context.OptionsLexical | Context.OptionsWebCompat],
     ['{ async function* f() {} let f }', Context.OptionsLexical],
     ['{ async function* f() {} let f }', Context.OptionsLexical | Context.OptionsWebCompat],
     ['{ async function* f() {}; var f; }', Context.OptionsLexical],
@@ -120,10 +127,12 @@ describe('Lexical - Block', () => {
     ['{ class f {} var f; }', Context.OptionsLexical],
     ['{ function* f() {} class f {}; }', Context.OptionsLexical],
     ['{ function* f() {} function f() {} }', Context.OptionsLexical],
+    ['{ function* f() {} function f() {} }', Context.OptionsLexical | Context.OptionsWebCompat],
     ['{ function *foo() {}; function *foo() {}; }', Context.OptionsLexical],
     ['{ function *foo() {}; function *bar() {}; function *foo() {}; }', Context.OptionsLexical],
     ['{ function* f() {} function* f() {} }', Context.OptionsLexical],
     ['{ function* f() {} function* f() {} }', Context.OptionsLexical | Context.Strict],
+    ['{ function* f() {} function* f() {} }', Context.OptionsLexical | Context.OptionsWebCompat],
     ['{ function* f() {} let f; }', Context.OptionsLexical],
     ['{ function* f() {} var f; }', Context.OptionsLexical],
     ['{ function f(){} function f(){} }', Context.OptionsLexical],
@@ -133,7 +142,6 @@ describe('Lexical - Block', () => {
     ['{ let f = 123; if (false) ; else function f() {} }', Context.OptionsLexical],
     ['{ let f = 123; if (false) ; else function f() {  }}', Context.OptionsLexical],
     ['try { throw null; } catch (f) { if (false) ; else function f() { return 123; } }', Context.OptionsLexical],
-    ['{ async function f() {} function f() {} }', Context.OptionsLexical],
     ['{ async function* f() {}; { var f; } }', Context.OptionsLexical],
     ['{ async function f() {}; { var f; } }', Context.OptionsLexical],
     ['{ class f {}; { var f; } }', Context.OptionsLexical],
@@ -202,7 +210,6 @@ describe('Lexical - Block', () => {
   for (const arg of [
     'function x() { { var f; var f } }',
     '{ { var f; } var f }',
-    '{ function f() {} async function* f() {} }',
     '{ function f() {} ; function f() {} }',
     '{ if (x) function f() {} ; function f() {} }',
     '{ var f = 123; if (true) function f(){} }',
@@ -220,8 +227,6 @@ describe('Lexical - Block', () => {
     '{ let foo = 1; { let foo = 2; } }',
     `function f() {}  var f;
     function f() {}  var f;`,
-    '{ async function f() {} async function* f() {} }',
-    '{ async function f() {} function f() {} }',
     '{ function f() { a = f; f = 123; b = f; return x; } }',
     '{ let f = 123; { function f() {  } } }',
     '{ let f = 123; if (false) ; else function f() {  } }',
@@ -230,18 +235,13 @@ describe('Lexical - Block', () => {
     `try { throw {}; } catch ({ f }) { switch (1) { default: function f() {  }} }
     try { throw {}; } catch ({ f }) { switch (1) { default: function f() {  }} }`,
     'let f = 123; switch (1) { default: function f() {  }  }',
-    '{ async function* f() {} async function f() {} }',
-    '{ function* f() {} function f() {} }',
     '{ let x; } var x',
-    '{ function* f() {} function* f() {} }',
     '{ var f; var f; }',
     `{ var f; var f; }
     { var f; var f; }`,
     'function f() {}  var f;',
     '{ function a(){} function a(){} }',
     '{ var a; { let a; } }',
-    '{ async function* f() {} function f() {} }',
-    '{ function* f() {} async function f() {} }',
     `{
       let result;
       let x = 1;
