@@ -107,6 +107,9 @@ describe('Lexical - Function', () => {
     ['(function() { { function* foo() {} function* foo() {} } })()', Context.OptionsLexical],
     ['(function() { { function* foo() {} function foo() {} } })()', Context.OptionsLexical],
     ['(function() { { function foo() {} function* foo() {} } })()', Context.OptionsLexical],
+    ['(function() { { function* foo() {} function* foo() {} } })()', Context.OptionsLexical | Context.OptionsWebCompat],
+    ['(function() { { function* foo() {} function foo() {} } })()', Context.OptionsLexical | Context.OptionsWebCompat],
+    ['(function() { { function foo() {} function* foo() {} } })()', Context.OptionsLexical | Context.OptionsWebCompat],
     ['(function() { { async function foo() {} async function foo() {} } })()', Context.OptionsLexical],
     ['function f(...rest, b){}', Context.OptionsLexical],
     ['let x; { var x; }', Context.OptionsLexical],
@@ -276,10 +279,6 @@ function a() {}`,
     'function foo([x]) { var x = 10;}',
     'async function af(x) { function x() { } }',
     'async function af(x) { var x; }',
-    '(function() { { function* foo() {} function* foo() {} } })()',
-    '(function() { { function* foo() {} function foo() {} } })()',
-    '(function() { { function foo() {} function* foo() {} } })()',
-    '(function() { { async function foo() {} async function foo() {} } })()',
     'function g() { var x = 1; { let x = 2; function g() { x; } g(); } }',
     'function f(one) { class x { } { class x { } function g() { one; x; } g() } } f()',
     'function *f(){} { function *f(){} }',
