@@ -48,6 +48,10 @@ describe('Lexical - Switch', () => {
     ['switch (0) { case 1: class f {} default: const f = 0 }', Context.OptionsLexical],
     ['switch (0) { case 1: async function* f() {} default: class f {} }', Context.OptionsLexical],
     ['switch (0) { case 1: async function f() {} default: function f() {} }', Context.OptionsLexical],
+    [
+      'switch (0) { case 1: async function f() {} default: function f() {} }',
+      Context.OptionsLexical | Context.OptionsWebCompat
+    ],
     ['switch (0) { case 1: async function* f() {} default: const f = 0; }', Context.OptionsLexical],
     ['switch (0) { case 1: async function* f() {} default: function f() {} }', Context.OptionsLexical],
     ['switch (0) { case 1: async function* f() {} default: function* f() {} }', Context.OptionsLexical],
@@ -176,9 +180,18 @@ describe('Lexical - Switch', () => {
     ['switch (x) { case c: function(){} function(){} }', Context.OptionsLexical],
     ['switch (x) { case c: async function f(){} async function f(){} }', Context.OptionsLexical],
     ['switch (x) { default: async function f(){} async function f(){} }', Context.OptionsLexical],
+    [
+      'switch (x) { default: async function f(){} async function f(){} }',
+      Context.OptionsLexical | Context.OptionsWebCompat
+    ],
     ['switch (x) { default: async function *f(){} async function *f(){} }', Context.OptionsLexical],
+    [
+      'switch (x) { default: async function *f(){} async function *f(){} }',
+      Context.OptionsLexical | Context.OptionsWebCompat
+    ],
     ['switch (x) { default: function f(){} function f(){} }', Context.OptionsLexical | Context.Strict],
-    ['switch (x) { default: function *f(){} function *f(){} }', Context.OptionsLexical]
+    ['switch (x) { default: function *f(){} function *f(){} }', Context.OptionsLexical],
+    ['switch (x) { default: function *f(){} function *f(){} }', Context.OptionsLexical | Context.OptionsWebCompat]
   ]);
 
   for (const arg of [
@@ -218,19 +231,13 @@ describe('Lexical - Switch', () => {
     `switch (0) { case 1: var f = 0; x; default: var {f} = x; } var {f} = f
     switch (0) { case 1: var f = 0; x; default: var {f} = x; } var {f} = f`,
     'switch (0) { case 1: let f = 0; x; default: let x; } var {f} = f',
-    'switch (0) { case 1: function f() {} default: async function f() {} }',
-    'switch (0) { case 1: async function f() {} default: function f() {} }',
     'switch (x) { case c: function f(){} function f(){} }',
-    'switch (x) { default: function *f(){} function *f(){} }',
-    'switch (x) { case c: function *f(){} function *f(){} }',
     'switch (0) { case 1: let f = 0; default: [f] }',
     `switch (0) { case 1: let f = 0; default: [f] }
     switch (0) { case 1: let f = 0; default: [f] }`,
-    'switch (x) { default: async function f(){} async function f(){} }',
     'switch (0) { default: let f; if (false) ; else function f() {  } }',
     'switch (0) { case 1: var f; default: var f; }',
     'switch (x) { case c: function f(){} function f(){} }',
-    'switch (x) { case c: async function *f(){} async function *f(){} }',
     `switch (0) { case 1: var f; default: var f; }
     switch (0) { case 1: var f; default: var f; }`,
     'for (let f of [0]) { switch (1) { case 1:function f() {  } }}'
