@@ -5,59 +5,69 @@ import { parseSource } from '../../../src/parser';
 
 describe('Next - Private methods', () => {
   fail('Private methods (fail)', [
-    ['class A { #a }', Context.OptionsWebCompat],
-    ['class A { #a }', Context.None],
-    ['a = { #ab() {} }', Context.OptionsNext],
-    ['class A { [{#ab() {}}]() {} }', Context.OptionsNext],
-    ['class A{ # a() {}}', Context.OptionsNext],
-    ['class C{ #method() { super(); } }', Context.OptionsNext],
-    ['var o = { #m() {} };', Context.OptionsNext],
-    ['var o = { async #m() {} };', Context.OptionsNext],
-    ['var o = { *#m() {} };', Context.OptionsNext],
-    ['class A { constructor = 4 }', Context.OptionsNext],
-    ['class A { #constructor = 4 }', Context.OptionsNext],
-    ['class A { a = () => super() }', Context.OptionsNext],
-    ['class A { # a }', Context.OptionsNext],
-    ['class A { #a; a() { this.# a } }', Context.OptionsNext],
-    ['class A {  #x; g = this.f; x = delete (g().#m); }', Context.OptionsNext],
-    ['class A {  #x; x = delete (this.#m); }', Context.OptionsNext],
-    ['class A {  #x; g = this.f; x = delete (g().#m); }', Context.OptionsNext],
-    ['class A {  #x; x = delete (g().#m); async *#m() {} }', Context.OptionsNext],
-    ['class A {  #x; x() { var g = this.f; delete g().#x; } }', Context.OptionsNext],
-    ['class A {  #x; x() { delete ((this.#m ));  } async #m() {} }', Context.OptionsNext],
-    ['class A { # x }', Context.OptionsNext],
-    ['class A { #\\u0000; }', Context.OptionsNext],
-    ['class A { * # m() {} }', Context.OptionsNext],
-    ['class A { # x; }', Context.OptionsNext],
-    ['class A { #x; m() { this.# x; }}', Context.OptionsNext],
-    ['class A { # m() {} }', Context.OptionsNext],
-    ['class A { static get # m() {} }', Context.OptionsNext],
-    ['class A { * method() { super(); } }', Context.OptionsNext],
-    ['class A { static async * prototype() {} }', Context.OptionsNext],
-    ['class A { static async #method() { super(); } }', Context.OptionsNext],
-    ['class A { method() { this.\\u0023field; } }', Context.OptionsNext],
-    ['class C { static #x = super(); }', Context.OptionsNext],
-    ['class C { async \\u0023m() {} } }', Context.OptionsNext],
-    ['class C { static async *#gen() { var await; } }', Context.OptionsNext],
-    ['class C { static async *#gen() { void await; } }', Context.OptionsNext],
-    ['class C { static async *#gen() { var yield; } }', Context.OptionsNext],
-    ['class C { static async *#gen() { void yield; } }', Context.OptionsNext],
-    ['class C { static async *#gen() { yield: ; } }', Context.OptionsNext],
-    ['class C { static async *#gen() {  void \\u0061wait; } }', Context.OptionsNext],
-    ['class C { async \\u0023m() {} } }', Context.OptionsNext],
-    ['class C { #method() { a() { foo().\\u0023field; } } }', Context.OptionsNext],
-    ['class C { #method() { \\u0023field; } }', Context.OptionsNext],
-    ['class C {  async \\u0023m() {} }', Context.OptionsNext],
-    ['var C = class { x = 42  *gen() {} }', Context.OptionsNext],
-    ['class Spaces { #  wrongSpaces() { return fail(); } }', Context.OptionsNext],
-    ['class C{ #method() { #[m] = 1} }', Context.OptionsNext],
-    ['class C{ #method() { #"p" = x } }', Context.OptionsNext],
-    ['class C{ #method() { super(); } }', Context.OptionsNext],
-    ['(class C extends Base { async #*a() { } })', Context.OptionsNext],
-    ['class C{ #method() { super(); } }', Context.OptionsNext],
-    ['class C{ #method() { super(); } }', Context.OptionsNext],
-    ['class C{ #method() { super(); } }', Context.OptionsNext],
-    ['class C { #x = () => arguments; }', Context.OptionsNext]
+    ['class A { #a b() {} }', Context.None],
+    ['class A { #a b }', Context.None],
+    ['class A { #a #b }', Context.None],
+    ['class A { a #b }', Context.None],
+    ['a = { #ab() {} }', Context.None],
+    ['class A { [{#ab() {}}]() {} }', Context.None],
+    ['class A{ # a() {}}', Context.None],
+    ['class C{ #method() { super(); } }', Context.None],
+    ['var o = { #m() {} };', Context.None],
+    ['var o = { async #m() {} };', Context.None],
+    ['var o = { *#m() {} };', Context.None],
+    ['class A { constructor = 4 }', Context.None],
+    ['class A { #constructor = 4 }', Context.None],
+    ['class A { a = () => super() }', Context.None],
+    ['class A { # a }', Context.None],
+    ['class A { #a; a() { this.# a } }', Context.None],
+    ['class A {  #x; g = this.f; x = delete (g().#m); }', Context.None],
+    ['class A {  #x; x = delete (this.#m); }', Context.None],
+    ['class A {  #x; g = this.f; x = delete (g().#m); }', Context.None],
+    ['class A {  #x; x = delete (g().#m); async *#m() {} }', Context.None],
+    ['class A {  #x; x() { var g = this.f; delete g().#x; } }', Context.None],
+    ['class A {  #x; x() { delete ((this.#m ));  } async #m() {} }', Context.None],
+    ['class A { # x }', Context.None],
+    ['class A { #\\u0000; }', Context.None],
+    ['class A { * # m() {} }', Context.None],
+    ['class A { # x; }', Context.None],
+    ['class A { #x; m() { this.# x; }}', Context.None],
+    ['class A { # m() {} }', Context.None],
+    ['class A { static get # m() {} }', Context.None],
+    ['class A { * method() { super(); } }', Context.None],
+    ['class A { static async * prototype() {} }', Context.None],
+    ['class A { static async #method() { super(); } }', Context.None],
+    ['class A { method() { this.\\u0023field; } }', Context.None],
+    ['class C { static #x = super(); }', Context.None],
+    ['class C { async \\u0023m() {} } }', Context.None],
+    ['class C { static async *#gen() { var await; } }', Context.None],
+    ['class C { static async *#gen() { void await; } }', Context.None],
+    ['class C { static async *#gen() { var yield; } }', Context.None],
+    ['class C { static async *#gen() { void yield; } }', Context.None],
+    ['class C { static async *#gen() { yield: ; } }', Context.None],
+    ['class C { static async *#gen() {  void \\u0061wait; } }', Context.None],
+    ['class C { async \\u0023m() {} } }', Context.None],
+    ['class C { #method() { a() { foo().\\u0023field; } } }', Context.None],
+    ['class C { #method() { \\u0023field; } }', Context.None],
+    ['class C {  async \\u0023m() {} }', Context.None],
+    ['var C = class { x = 42  *gen() {} }', Context.None],
+    ['class Spaces { #  wrongSpaces() { return fail(); } }', Context.None],
+    ['class C{ #method() { #[m] = 1} }', Context.None],
+    ['class C{ #method() { #"p" = x } }', Context.None],
+    ['class C{ #method() { super(); } }', Context.None],
+    ['(class C extends Base { async #*a() { } })', Context.None],
+    ['class C{ #method() { super(); } }', Context.None],
+    ['class C{ #method() { super(); } }', Context.None],
+    ['class C{ #method() { super(); } }', Context.None],
+    ['class C { #x = () => arguments; }', Context.None],
+    ['class A { #x; #x }', Context.OptionsLexical],
+    ['class A { get #x() {} get #x() {} }', Context.OptionsLexical],
+    ['class A { set #x(v) {} set #x(v) {} }', Context.OptionsLexical],
+    ['class A { static get #x() {} static get #x() {} }', Context.OptionsLexical],
+    ['class A { static set #x(v) {} static set #x(v) {} }', Context.OptionsLexical],
+    ['class A { #x; fn() { return this.#y } }', Context.OptionsLexical],
+    ['class A { #x } function fn() { return this.#y }', Context.OptionsLexical],
+    ['class A extends B { #x() {} method() { super.#x() }  }', Context.OptionsLexical]
   ]);
 
   for (const arg of [
@@ -110,25 +120,25 @@ describe('Next - Private methods', () => {
   ]) {
     it(`class C { ${arg} }`, () => {
       t.throws(() => {
-        parseSource(`class C { ${arg} }`, undefined, Context.OptionsNext);
+        parseSource(`class C { ${arg} }`, undefined, Context.None);
       });
     });
 
     it(`class C extends Base { ${arg} }`, () => {
       t.throws(() => {
-        parseSource(`class C extends Base { ${arg} }`, undefined, Context.OptionsNext);
+        parseSource(`class C extends Base { ${arg} }`, undefined, Context.None);
       });
     });
 
     it(`(class C { ${arg} })`, () => {
       t.throws(() => {
-        parseSource(`(class C { ${arg} })`, undefined, Context.OptionsNext);
+        parseSource(`(class C { ${arg} })`, undefined, Context.None);
       });
     });
 
     it(`(class C extends Base { ${arg} })`, () => {
       t.throws(() => {
-        parseSource(`(class C extends Base { ${arg} })`, undefined, Context.OptionsNext);
+        parseSource(`(class C extends Base { ${arg} })`, undefined, Context.None);
       });
     });
   }
@@ -192,31 +202,40 @@ describe('Next - Private methods', () => {
     'foo() { this.#m, (() => this)().#m }',
     'foo() { this.#m, (() => this)().#m }',
     'foo() { this.#m, (() => this)().#m }',
-    'method() { super.#x(); }',
     '#method() { super.x(); }',
-    '#method() { super.#x(); }'
+    '#\\u{61}',
+    '#\\u{61}\n',
+    '#\\u{61}bc;',
+    '#\\u{61};\n',
+    '#\\u{61} = 2',
+    '#\\u{61} = 2;\n',
+    'static #\\u{61}',
+    'static #\\u{61}\n',
+    'static #\\u{61};',
+    'static #\\u{61}bc=2',
+    'static #\\u{61} = 2;\n'
   ]) {
     it(`class C { ${arg} }`, () => {
       t.doesNotThrow(() => {
-        parseSource(`class C { ${arg} }`, undefined, Context.OptionsNext);
+        parseSource(`class C { ${arg} }`, undefined, Context.None);
       });
     });
 
     it(`class C extends Base { ${arg} }`, () => {
       t.doesNotThrow(() => {
-        parseSource(`class C extends Base { ${arg} }`, undefined, Context.OptionsNext);
+        parseSource(`class C extends Base { ${arg} }`, undefined, Context.None);
       });
     });
 
     it(`(class C { ${arg} })`, () => {
       t.doesNotThrow(() => {
-        parseSource(`(class C { ${arg} })`, undefined, Context.OptionsNext);
+        parseSource(`(class C { ${arg} })`, undefined, Context.None);
       });
     });
 
     it(`(class C extends Base { ${arg} })`, () => {
       t.doesNotThrow(() => {
-        parseSource(`(class C extends Base { ${arg} })`, undefined, Context.OptionsNext);
+        parseSource(`(class C extends Base { ${arg} })`, undefined, Context.None);
       });
     });
   }
@@ -235,7 +254,6 @@ describe('Next - Private methods', () => {
     'var C = class { static async m() { return 42; } #℘_; ℘(value) { this.#℘_ = value; return this.#℘; }}',
     'var C = class { static async m() { return 42; } static * #$(value) { yield * value; }  static * #_(value) {  yield * value; } }',
     'var C = class { static m() { return 42; } static #$ = 1; static #_ = 1; static _() { return this.#_; } }',
-    'var C = class { var C = class { [obj1] = 42; [obj2] = 43; [obj3] = 44; } }',
     'var C = class { static m() { return 42; } static #$ = 1; static #_ = 1; static _() { return this.#_; } }',
     '{ class C { #a() { class B { #a() {  } } new B; } } new C; }',
     `{
@@ -261,14 +279,14 @@ describe('Next - Private methods', () => {
   ]) {
     it(`${arg}`, () => {
       t.doesNotThrow(() => {
-        parseSource(`${arg}`, undefined, Context.OptionsNext);
+        parseSource(`${arg}`, undefined, Context.None);
       });
     });
   }
   pass('Next - Private methods (pass)', [
     [
       `class A { #key; }`,
-      Context.OptionsNext,
+      Context.None,
       {
         body: [
           {
@@ -276,7 +294,7 @@ describe('Next - Private methods', () => {
               body: [
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: 'key',
                     type: 'PrivateIdentifier'
@@ -288,7 +306,7 @@ describe('Next - Private methods', () => {
               ],
               type: 'ClassBody'
             },
-            decorators: [],
+
             id: {
               name: 'A',
               type: 'Identifier'
@@ -303,7 +321,7 @@ describe('Next - Private methods', () => {
     ],
     [
       `class A { static async #_(value) { return await value;} }`,
-      Context.OptionsNext,
+      Context.None,
       {
         body: [
           {
@@ -315,7 +333,7 @@ describe('Next - Private methods', () => {
                     name: '_',
                     type: 'PrivateIdentifier'
                   },
-                  decorators: [],
+
                   kind: 'method',
                   static: true,
                   type: 'MethodDefinition',
@@ -350,7 +368,7 @@ describe('Next - Private methods', () => {
               ],
               type: 'ClassBody'
             },
-            decorators: [],
+
             id: {
               name: 'A',
               type: 'Identifier'
@@ -365,14 +383,14 @@ describe('Next - Private methods', () => {
     ],
     [
       `class A { #a; #b; }`,
-      Context.OptionsNext | Context.OptionsRanges,
+      Context.OptionsRanges,
       {
         type: 'Program',
         sourceType: 'script',
         body: [
           {
             type: 'ClassDeclaration',
-            decorators: [],
+
             id: {
               type: 'Identifier',
               name: 'A',
@@ -386,7 +404,7 @@ describe('Next - Private methods', () => {
               body: [
                 {
                   type: 'PropertyDefinition',
-                  decorators: [],
+
                   key: {
                     type: 'PrivateIdentifier',
                     name: 'a',
@@ -398,12 +416,12 @@ describe('Next - Private methods', () => {
                   computed: false,
                   static: false,
                   start: 10,
-                  end: 12,
-                  range: [10, 12]
+                  end: 13,
+                  range: [10, 13]
                 },
                 {
                   type: 'PropertyDefinition',
-                  decorators: [],
+
                   key: {
                     type: 'PrivateIdentifier',
                     name: 'b',
@@ -415,8 +433,8 @@ describe('Next - Private methods', () => {
                   computed: false,
                   static: false,
                   start: 14,
-                  end: 16,
-                  range: [14, 16]
+                  end: 17,
+                  range: [14, 17]
                 }
               ],
               start: 8,
@@ -435,14 +453,14 @@ describe('Next - Private methods', () => {
     ],
     [
       `class A { #yield = b[c]; }`,
-      Context.OptionsNext | Context.OptionsRanges,
+      Context.OptionsRanges,
       {
         type: 'Program',
         sourceType: 'script',
         body: [
           {
             type: 'ClassDeclaration',
-            decorators: [],
+
             id: {
               type: 'Identifier',
               name: 'A',
@@ -456,7 +474,7 @@ describe('Next - Private methods', () => {
               body: [
                 {
                   type: 'PropertyDefinition',
-                  decorators: [],
+
                   key: {
                     type: 'PrivateIdentifier',
                     name: 'yield',
@@ -488,8 +506,8 @@ describe('Next - Private methods', () => {
                   computed: false,
                   static: false,
                   start: 10,
-                  end: 23,
-                  range: [10, 23]
+                  end: 24,
+                  range: [10, 24]
                 }
               ],
               start: 8,
@@ -508,14 +526,14 @@ describe('Next - Private methods', () => {
     ],
     [
       `class A { #yield = foo + bar; }`,
-      Context.OptionsNext | Context.OptionsRanges,
+      Context.OptionsRanges,
       {
         type: 'Program',
         sourceType: 'script',
         body: [
           {
             type: 'ClassDeclaration',
-            decorators: [],
+
             id: {
               type: 'Identifier',
               name: 'A',
@@ -529,7 +547,7 @@ describe('Next - Private methods', () => {
               body: [
                 {
                   type: 'PropertyDefinition',
-                  decorators: [],
+
                   key: {
                     type: 'PrivateIdentifier',
                     name: 'yield',
@@ -561,8 +579,8 @@ describe('Next - Private methods', () => {
                   computed: false,
                   static: false,
                   start: 10,
-                  end: 28,
-                  range: [10, 28]
+                  end: 29,
+                  range: [10, 29]
                 }
               ],
               start: 8,
@@ -581,14 +599,14 @@ describe('Next - Private methods', () => {
     ],
     [
       `class A { #yield; }`,
-      Context.OptionsNext | Context.OptionsRanges,
+      Context.OptionsRanges,
       {
         type: 'Program',
         sourceType: 'script',
         body: [
           {
             type: 'ClassDeclaration',
-            decorators: [],
+
             id: {
               type: 'Identifier',
               name: 'A',
@@ -602,7 +620,7 @@ describe('Next - Private methods', () => {
               body: [
                 {
                   type: 'PropertyDefinition',
-                  decorators: [],
+
                   key: {
                     type: 'PrivateIdentifier',
                     name: 'yield',
@@ -614,8 +632,8 @@ describe('Next - Private methods', () => {
                   computed: false,
                   static: false,
                   start: 10,
-                  end: 16,
-                  range: [10, 16]
+                  end: 17,
+                  range: [10, 17]
                 }
               ],
               start: 8,
@@ -634,7 +652,7 @@ describe('Next - Private methods', () => {
     ],
     [
       `class C { static async *#gen() { yield { ...yield,  y: 1, ...yield yield, }; } static get gen() { return this.#gen; } }`,
-      Context.OptionsNext,
+      Context.None,
       {
         body: [
           {
@@ -648,7 +666,7 @@ describe('Next - Private methods', () => {
                   },
                   kind: 'method',
                   static: true,
-                  decorators: [],
+
                   type: 'MethodDefinition',
                   value: {
                     async: true,
@@ -712,7 +730,7 @@ describe('Next - Private methods', () => {
                 },
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: 'gen',
                     type: 'Identifier'
@@ -750,7 +768,7 @@ describe('Next - Private methods', () => {
               ],
               type: 'ClassBody'
             },
-            decorators: [],
+
             id: {
               name: 'C',
               type: 'Identifier'
@@ -765,7 +783,7 @@ describe('Next - Private methods', () => {
     ],
     [
       `class C { static *#gen() { yield [...yield yield]; } static get gen() { return this.#gen; } }`,
-      Context.OptionsNext,
+      Context.None,
       {
         body: [
           {
@@ -779,7 +797,7 @@ describe('Next - Private methods', () => {
                   },
                   kind: 'method',
                   static: true,
-                  decorators: [],
+
                   type: 'MethodDefinition',
                   value: {
                     async: false,
@@ -820,7 +838,7 @@ describe('Next - Private methods', () => {
                 },
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: 'gen',
                     type: 'Identifier'
@@ -858,7 +876,7 @@ describe('Next - Private methods', () => {
               ],
               type: 'ClassBody'
             },
-            decorators: [],
+
             id: {
               name: 'C',
               type: 'Identifier'
@@ -873,7 +891,7 @@ describe('Next - Private methods', () => {
     ],
     [
       `class C { get #℘() {} set #℘(x) {} a() { return this.#℘; } b(value) { this.#℘ = x; } };`,
-      Context.OptionsNext,
+      Context.None,
       {
         body: [
           {
@@ -887,7 +905,7 @@ describe('Next - Private methods', () => {
                   },
                   kind: 'get',
                   static: false,
-                  decorators: [],
+
                   type: 'MethodDefinition',
                   value: {
                     async: false,
@@ -909,7 +927,7 @@ describe('Next - Private methods', () => {
                   },
                   kind: 'set',
                   static: false,
-                  decorators: [],
+
                   type: 'MethodDefinition',
                   value: {
                     async: false,
@@ -930,7 +948,7 @@ describe('Next - Private methods', () => {
                 },
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: 'a',
                     type: 'Identifier'
@@ -967,7 +985,7 @@ describe('Next - Private methods', () => {
                 },
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: 'b',
                     type: 'Identifier'
@@ -1018,7 +1036,7 @@ describe('Next - Private methods', () => {
               ],
               type: 'ClassBody'
             },
-            decorators: [],
+
             id: {
               name: 'C',
               type: 'Identifier'
@@ -1036,14 +1054,14 @@ describe('Next - Private methods', () => {
     ],
     [
       `class C { #m() { return 42; } get ref() { return this.#m; } }`,
-      Context.OptionsNext,
+      Context.None,
       {
         type: 'Program',
         sourceType: 'script',
         body: [
           {
             type: 'ClassDeclaration',
-            decorators: [],
+
             id: {
               type: 'Identifier',
               name: 'C'
@@ -1057,7 +1075,7 @@ describe('Next - Private methods', () => {
                   kind: 'method',
                   static: false,
                   computed: false,
-                  decorators: [],
+
                   key: {
                     type: 'PrivateIdentifier',
                     name: 'm'
@@ -1087,7 +1105,7 @@ describe('Next - Private methods', () => {
                   kind: 'get',
                   static: false,
                   computed: false,
-                  decorators: [],
+
                   key: {
                     type: 'Identifier',
                     name: 'ref'
@@ -1127,7 +1145,7 @@ describe('Next - Private methods', () => {
     ],
     [
       `class A { #foo = bar; }`,
-      Context.OptionsNext,
+      Context.None,
       {
         body: [
           {
@@ -1135,7 +1153,7 @@ describe('Next - Private methods', () => {
               body: [
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: 'foo',
                     type: 'PrivateIdentifier'
@@ -1150,7 +1168,7 @@ describe('Next - Private methods', () => {
               ],
               type: 'ClassBody'
             },
-            decorators: [],
+
             id: {
               name: 'A',
               type: 'Identifier'
@@ -1211,7 +1229,7 @@ describe('Next - Private methods', () => {
           this.publicFieldValue = -(this.publicFieldValue ** this.publicFieldValue);
         }
       }`,
-      Context.OptionsNext,
+      Context.None,
       {
         type: 'Program',
         sourceType: 'script',
@@ -1223,7 +1241,7 @@ describe('Next - Private methods', () => {
               name: 'Cl'
             },
             superClass: null,
-            decorators: [],
+
             body: {
               type: 'ClassBody',
               body: [
@@ -1238,8 +1256,7 @@ describe('Next - Private methods', () => {
                     value: 'top secret string'
                   },
                   static: false,
-                  computed: false,
-                  decorators: []
+                  computed: false
                 },
                 {
                   type: 'MethodDefinition',
@@ -1250,7 +1267,7 @@ describe('Next - Private methods', () => {
                     type: 'Identifier',
                     name: 'constructor'
                   },
-                  decorators: [],
+
                   value: {
                     type: 'FunctionExpression',
                     params: [],
@@ -1291,7 +1308,7 @@ describe('Next - Private methods', () => {
                   kind: 'get',
                   static: false,
                   computed: false,
-                  decorators: [],
+
                   key: {
                     type: 'PrivateIdentifier',
                     name: 'privateFieldValue'
@@ -1328,7 +1345,7 @@ describe('Next - Private methods', () => {
                   kind: 'set',
                   static: false,
                   computed: false,
-                  decorators: [],
+
                   key: {
                     type: 'PrivateIdentifier',
                     name: 'privateFieldValue'
@@ -1382,7 +1399,7 @@ describe('Next - Private methods', () => {
                     type: 'Identifier',
                     name: 'publicGetPrivateField'
                   },
-                  decorators: [],
+
                   value: {
                     type: 'FunctionExpression',
                     params: [],
@@ -1419,7 +1436,7 @@ describe('Next - Private methods', () => {
                     type: 'Identifier',
                     name: 'publicSetPrivateField'
                   },
-                  decorators: [],
+
                   value: {
                     type: 'FunctionExpression',
                     params: [
@@ -1469,7 +1486,7 @@ describe('Next - Private methods', () => {
                     type: 'Identifier',
                     name: 'publicFieldValue'
                   },
-                  decorators: [],
+
                   value: {
                     type: 'FunctionExpression',
                     params: [],
@@ -1506,7 +1523,7 @@ describe('Next - Private methods', () => {
                     type: 'Identifier',
                     name: 'publicFieldValue'
                   },
-                  decorators: [],
+
                   value: {
                     type: 'FunctionExpression',
                     params: [
@@ -1556,7 +1573,7 @@ describe('Next - Private methods', () => {
                     type: 'Identifier',
                     name: 'testUpdates'
                   },
-                  decorators: [],
+
                   value: {
                     type: 'FunctionExpression',
                     params: [],
@@ -1880,7 +1897,7 @@ describe('Next - Private methods', () => {
             this.#privateFieldValue = 1;
           }
         }`,
-      Context.OptionsNext,
+      Context.None,
       {
         type: 'Program',
         sourceType: 'script',
@@ -1892,7 +1909,7 @@ describe('Next - Private methods', () => {
               name: 'Cl'
             },
             superClass: null,
-            decorators: [],
+
             body: {
               type: 'ClassBody',
               body: [
@@ -1907,15 +1924,14 @@ describe('Next - Private methods', () => {
                     value: 0
                   },
                   static: false,
-                  computed: false,
-                  decorators: []
+                  computed: false
                 },
                 {
                   type: 'MethodDefinition',
                   kind: 'get',
                   static: false,
                   computed: false,
-                  decorators: [],
+
                   key: {
                     type: 'PrivateIdentifier',
                     name: 'privateFieldValue'
@@ -1956,7 +1972,7 @@ describe('Next - Private methods', () => {
                     type: 'Identifier',
                     name: 'constructor'
                   },
-                  decorators: [],
+
                   value: {
                     type: 'FunctionExpression',
                     params: [],
@@ -2022,7 +2038,7 @@ describe('Next - Private methods', () => {
               this.#privateFieldValue = newValue;
             }
           } `,
-      Context.OptionsNext,
+      Context.None,
       {
         type: 'Program',
         sourceType: 'script',
@@ -2034,7 +2050,7 @@ describe('Next - Private methods', () => {
               name: 'Cl'
             },
             superClass: null,
-            decorators: [],
+
             body: {
               type: 'ClassBody',
               body: [
@@ -2049,8 +2065,7 @@ describe('Next - Private methods', () => {
                     value: 'top secret string'
                   },
                   static: false,
-                  computed: false,
-                  decorators: []
+                  computed: false
                 },
                 {
                   type: 'MethodDefinition',
@@ -2061,7 +2076,7 @@ describe('Next - Private methods', () => {
                     type: 'Identifier',
                     name: 'constructor'
                   },
-                  decorators: [],
+
                   value: {
                     type: 'FunctionExpression',
                     params: [],
@@ -2102,7 +2117,7 @@ describe('Next - Private methods', () => {
                   kind: 'get',
                   static: false,
                   computed: false,
-                  decorators: [],
+
                   key: {
                     type: 'PrivateIdentifier',
                     name: 'privateFieldValue'
@@ -2139,7 +2154,7 @@ describe('Next - Private methods', () => {
                   kind: 'set',
                   static: false,
                   computed: false,
-                  decorators: [],
+
                   key: {
                     type: 'PrivateIdentifier',
                     name: 'privateFieldValue'
@@ -2193,7 +2208,7 @@ describe('Next - Private methods', () => {
                     type: 'Identifier',
                     name: 'publicGetPrivateField'
                   },
-                  decorators: [],
+
                   value: {
                     type: 'FunctionExpression',
                     params: [],
@@ -2230,7 +2245,7 @@ describe('Next - Private methods', () => {
                     type: 'Identifier',
                     name: 'publicSetPrivateField'
                   },
-                  decorators: [],
+
                   value: {
                     type: 'FunctionExpression',
                     params: [
@@ -2279,14 +2294,14 @@ describe('Next - Private methods', () => {
     ],
     [
       `class A { #key() {} }`,
-      Context.OptionsNext,
+      Context.None,
       {
         type: 'Program',
         sourceType: 'script',
         body: [
           {
             type: 'ClassDeclaration',
-            decorators: [],
+
             id: {
               type: 'Identifier',
               name: 'A'
@@ -2300,7 +2315,7 @@ describe('Next - Private methods', () => {
                   kind: 'method',
                   static: false,
                   computed: false,
-                  decorators: [],
+
                   key: {
                     type: 'PrivateIdentifier',
                     name: 'key'
@@ -2325,7 +2340,7 @@ describe('Next - Private methods', () => {
     ],
     [
       `class A { #yield\n = 0; }`,
-      Context.OptionsNext,
+      Context.None,
       {
         body: [
           {
@@ -2333,7 +2348,7 @@ describe('Next - Private methods', () => {
               body: [
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: 'yield',
                     type: 'PrivateIdentifier'
@@ -2348,7 +2363,7 @@ describe('Next - Private methods', () => {
               ],
               type: 'ClassBody'
             },
-            decorators: [],
+
             id: {
               name: 'A',
               type: 'Identifier'
@@ -2363,14 +2378,14 @@ describe('Next - Private methods', () => {
     ],
     [
       `class A { #foo() { #bar } }`,
-      Context.OptionsNext,
+      Context.None,
       {
         type: 'Program',
         sourceType: 'script',
         body: [
           {
             type: 'ClassDeclaration',
-            decorators: [],
+
             id: {
               type: 'Identifier',
               name: 'A'
@@ -2382,7 +2397,7 @@ describe('Next - Private methods', () => {
                 {
                   type: 'MethodDefinition',
                   kind: 'method',
-                  decorators: [],
+
                   static: false,
                   computed: false,
                   key: {
@@ -2417,7 +2432,7 @@ describe('Next - Private methods', () => {
     ],
     [
       `class A { static #key; }`,
-      Context.OptionsNext,
+      Context.None,
       {
         body: [
           {
@@ -2425,7 +2440,7 @@ describe('Next - Private methods', () => {
               body: [
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: 'key',
                     type: 'PrivateIdentifier'
@@ -2437,7 +2452,7 @@ describe('Next - Private methods', () => {
               ],
               type: 'ClassBody'
             },
-            decorators: [],
+
             id: {
               name: 'A',
               type: 'Identifier'
@@ -2452,7 +2467,7 @@ describe('Next - Private methods', () => {
     ],
     [
       `class A { static #foo(bar) {} }`,
-      Context.OptionsNext,
+      Context.None,
       {
         body: [
           {
@@ -2460,7 +2475,7 @@ describe('Next - Private methods', () => {
               body: [
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: 'foo',
                     type: 'PrivateIdentifier'
@@ -2488,7 +2503,7 @@ describe('Next - Private methods', () => {
               ],
               type: 'ClassBody'
             },
-            decorators: [],
+
             id: {
               name: 'A',
               type: 'Identifier'
@@ -2503,7 +2518,7 @@ describe('Next - Private methods', () => {
     ],
     [
       `class A { m() {} #a; }`,
-      Context.OptionsNext,
+      Context.None,
       {
         body: [
           {
@@ -2511,7 +2526,7 @@ describe('Next - Private methods', () => {
               body: [
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: 'm',
                     type: 'Identifier'
@@ -2533,7 +2548,7 @@ describe('Next - Private methods', () => {
                 },
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: 'a',
                     type: 'PrivateIdentifier'
@@ -2545,7 +2560,7 @@ describe('Next - Private methods', () => {
               ],
               type: 'ClassBody'
             },
-            decorators: [],
+
             id: {
               name: 'A',
               type: 'Identifier'
@@ -2560,7 +2575,7 @@ describe('Next - Private methods', () => {
     ],
     [
       `class A {  #a; m() {} }`,
-      Context.OptionsNext,
+      Context.None,
       {
         body: [
           {
@@ -2568,7 +2583,7 @@ describe('Next - Private methods', () => {
               body: [
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: 'a',
                     type: 'PrivateIdentifier'
@@ -2579,7 +2594,7 @@ describe('Next - Private methods', () => {
                 },
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: 'm',
                     type: 'Identifier'
@@ -2602,7 +2617,7 @@ describe('Next - Private methods', () => {
               ],
               type: 'ClassBody'
             },
-            decorators: [],
+
             id: {
               name: 'A',
               type: 'Identifier'
@@ -2617,7 +2632,7 @@ describe('Next - Private methods', () => {
     ],
     [
       `class A {  #a; m() {} #b; }`,
-      Context.OptionsNext,
+      Context.None,
       {
         body: [
           {
@@ -2625,7 +2640,7 @@ describe('Next - Private methods', () => {
               body: [
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: 'a',
                     type: 'PrivateIdentifier'
@@ -2636,7 +2651,7 @@ describe('Next - Private methods', () => {
                 },
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: 'm',
                     type: 'Identifier'
@@ -2658,7 +2673,7 @@ describe('Next - Private methods', () => {
                 },
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: 'b',
                     type: 'PrivateIdentifier'
@@ -2670,7 +2685,7 @@ describe('Next - Private methods', () => {
               ],
               type: 'ClassBody'
             },
-            decorators: [],
+
             id: {
               name: 'A',
               type: 'Identifier'
@@ -2685,7 +2700,7 @@ describe('Next - Private methods', () => {
     ],
     [
       `class A { m() { return 42; } #a;  #__;  #NJ_;  #℘_ ; }`,
-      Context.OptionsNext,
+      Context.None,
       {
         body: [
           {
@@ -2693,7 +2708,7 @@ describe('Next - Private methods', () => {
               body: [
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: 'm',
                     type: 'Identifier'
@@ -2723,7 +2738,7 @@ describe('Next - Private methods', () => {
                 },
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: 'a',
                     type: 'PrivateIdentifier'
@@ -2734,7 +2749,7 @@ describe('Next - Private methods', () => {
                 },
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: '__',
                     type: 'PrivateIdentifier'
@@ -2745,7 +2760,7 @@ describe('Next - Private methods', () => {
                 },
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: 'NJ_',
                     type: 'PrivateIdentifier'
@@ -2756,7 +2771,7 @@ describe('Next - Private methods', () => {
                 },
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: '℘_',
                     type: 'PrivateIdentifier'
@@ -2768,7 +2783,7 @@ describe('Next - Private methods', () => {
               ],
               type: 'ClassBody'
             },
-            decorators: [],
+
             id: {
               name: 'A',
               type: 'Identifier'
@@ -2785,14 +2800,14 @@ describe('Next - Private methods', () => {
       `class A { #foo = () => 'bar';  method() {
         return this.#foo();
       } }`,
-      Context.OptionsNext,
+      Context.None,
       {
         type: 'Program',
         sourceType: 'script',
         body: [
           {
             type: 'ClassDeclaration',
-            decorators: [],
+
             id: {
               type: 'Identifier',
               name: 'A'
@@ -2803,7 +2818,7 @@ describe('Next - Private methods', () => {
               body: [
                 {
                   type: 'PropertyDefinition',
-                  decorators: [],
+
                   key: {
                     type: 'PrivateIdentifier',
                     name: 'foo'
@@ -2826,7 +2841,7 @@ describe('Next - Private methods', () => {
                   kind: 'method',
                   static: false,
                   computed: false,
-                  decorators: [],
+
                   key: {
                     type: 'Identifier',
                     name: 'method'
@@ -2870,7 +2885,7 @@ describe('Next - Private methods', () => {
     ],
     [
       `class B { #x = 0; #y = 1; }`,
-      Context.OptionsNext,
+      Context.None,
       {
         type: 'Program',
         sourceType: 'script',
@@ -2882,7 +2897,7 @@ describe('Next - Private methods', () => {
               name: 'B'
             },
             superClass: null,
-            decorators: [],
+
             body: {
               type: 'ClassBody',
               body: [
@@ -2897,8 +2912,7 @@ describe('Next - Private methods', () => {
                     value: 0
                   },
                   static: false,
-                  computed: false,
-                  decorators: []
+                  computed: false
                 },
                 {
                   type: 'PropertyDefinition',
@@ -2911,8 +2925,7 @@ describe('Next - Private methods', () => {
                     value: 1
                   },
                   static: false,
-                  computed: false,
-                  decorators: []
+                  computed: false
                 }
               ]
             }
@@ -2925,7 +2938,7 @@ describe('Next - Private methods', () => {
           static #x;
           static #y = 1;
         }`,
-      Context.OptionsNext,
+      Context.None,
       {
         type: 'Program',
         sourceType: 'script',
@@ -2937,7 +2950,7 @@ describe('Next - Private methods', () => {
               name: 'A'
             },
             superClass: null,
-            decorators: [],
+
             body: {
               type: 'ClassBody',
               body: [
@@ -2949,8 +2962,7 @@ describe('Next - Private methods', () => {
                   },
                   value: null,
                   static: true,
-                  computed: false,
-                  decorators: []
+                  computed: false
                 },
                 {
                   type: 'PropertyDefinition',
@@ -2963,8 +2975,7 @@ describe('Next - Private methods', () => {
                     value: 1
                   },
                   static: true,
-                  computed: false,
-                  decorators: []
+                  computed: false
                 }
               ]
             }
@@ -2974,7 +2985,7 @@ describe('Next - Private methods', () => {
     ],
     [
       `class A {  #m = async function() { return 'foo'; };  method() { return this.#m(); } }`,
-      Context.OptionsNext,
+      Context.None,
       {
         body: [
           {
@@ -2982,7 +2993,7 @@ describe('Next - Private methods', () => {
               body: [
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: 'm',
                     type: 'PrivateIdentifier'
@@ -3012,7 +3023,7 @@ describe('Next - Private methods', () => {
                 },
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: 'method',
                     type: 'Identifier'
@@ -3054,7 +3065,7 @@ describe('Next - Private methods', () => {
               ],
               type: 'ClassBody'
             },
-            decorators: [],
+
             id: {
               name: 'A',
               type: 'Identifier'
@@ -3069,7 +3080,7 @@ describe('Next - Private methods', () => {
     ],
     [
       `class A { method() { return this.#m(); } #m = function () { return 'foo'; };  }`,
-      Context.OptionsNext,
+      Context.None,
       {
         body: [
           {
@@ -3077,7 +3088,7 @@ describe('Next - Private methods', () => {
               body: [
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: 'method',
                     type: 'Identifier'
@@ -3118,7 +3129,7 @@ describe('Next - Private methods', () => {
                 },
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: 'm',
                     type: 'PrivateIdentifier'
@@ -3149,7 +3160,7 @@ describe('Next - Private methods', () => {
               ],
               type: 'ClassBody'
             },
-            decorators: [],
+
             id: {
               name: 'A',
               type: 'Identifier'
@@ -3164,7 +3175,7 @@ describe('Next - Private methods', () => {
     ],
     [
       `class A {  #m() { return 42; } get bGetter() { return this.#b; } #b = this.#m(); get ref() { return this.#m; } constructor() {} }`,
-      Context.OptionsNext,
+      Context.None,
       {
         body: [
           {
@@ -3172,7 +3183,7 @@ describe('Next - Private methods', () => {
               body: [
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: 'm',
                     type: 'PrivateIdentifier'
@@ -3202,7 +3213,7 @@ describe('Next - Private methods', () => {
                 },
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: 'bGetter',
                     type: 'Identifier'
@@ -3239,7 +3250,7 @@ describe('Next - Private methods', () => {
                 },
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: 'b',
                     type: 'PrivateIdentifier'
@@ -3264,7 +3275,7 @@ describe('Next - Private methods', () => {
                 },
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: 'ref',
                     type: 'Identifier'
@@ -3301,7 +3312,7 @@ describe('Next - Private methods', () => {
                 },
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: 'constructor',
                     type: 'Identifier'
@@ -3324,7 +3335,7 @@ describe('Next - Private methods', () => {
               ],
               type: 'ClassBody'
             },
-            decorators: [],
+
             id: {
               name: 'A',
               type: 'Identifier'
@@ -3339,7 +3350,7 @@ describe('Next - Private methods', () => {
     ],
     [
       `class A { #$_; #℘_; }`,
-      Context.OptionsNext,
+      Context.None,
       {
         body: [
           {
@@ -3347,7 +3358,7 @@ describe('Next - Private methods', () => {
               body: [
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: '$_',
                     type: 'PrivateIdentifier'
@@ -3358,7 +3369,7 @@ describe('Next - Private methods', () => {
                 },
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: '℘_',
                     type: 'PrivateIdentifier'
@@ -3370,7 +3381,7 @@ describe('Next - Private methods', () => {
               ],
               type: 'ClassBody'
             },
-            decorators: [],
+
             id: {
               name: 'A',
               type: 'Identifier'
@@ -3385,7 +3396,7 @@ describe('Next - Private methods', () => {
     ],
     [
       `class A { $(value) { this.#$_ = value; return this.#$; } }`,
-      Context.OptionsNext,
+      Context.None,
       {
         body: [
           {
@@ -3393,7 +3404,7 @@ describe('Next - Private methods', () => {
               body: [
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: '$',
                     type: 'Identifier'
@@ -3458,7 +3469,7 @@ describe('Next - Private methods', () => {
               ],
               type: 'ClassBody'
             },
-            decorators: [],
+
             id: {
               name: 'A',
               type: 'Identifier'
@@ -3479,7 +3490,7 @@ describe('Next - Private methods', () => {
           this.#__ = value;
         }
       }`,
-      Context.OptionsNext,
+      Context.None,
       {
         type: 'Program',
         sourceType: 'script',
@@ -3494,7 +3505,7 @@ describe('Next - Private methods', () => {
                   type: 'ClassExpression',
                   id: null,
                   superClass: null,
-                  decorators: [],
+
                   body: {
                     type: 'ClassBody',
                     body: [
@@ -3507,7 +3518,7 @@ describe('Next - Private methods', () => {
                           type: 'Identifier',
                           name: 'm'
                         },
-                        decorators: [],
+
                         value: {
                           type: 'FunctionExpression',
                           params: [],
@@ -3536,8 +3547,7 @@ describe('Next - Private methods', () => {
                         },
                         value: null,
                         static: false,
-                        computed: false,
-                        decorators: []
+                        computed: false
                       },
                       {
                         type: 'PropertyDefinition',
@@ -3547,8 +3557,7 @@ describe('Next - Private methods', () => {
                         },
                         value: null,
                         static: false,
-                        computed: false,
-                        decorators: []
+                        computed: false
                       },
                       {
                         type: 'PropertyDefinition',
@@ -3558,15 +3567,14 @@ describe('Next - Private methods', () => {
                         },
                         value: null,
                         static: false,
-                        computed: false,
-                        decorators: []
+                        computed: false
                       },
                       {
                         type: 'MethodDefinition',
                         kind: 'set',
                         static: false,
                         computed: false,
-                        decorators: [],
+
                         key: {
                           type: 'PrivateIdentifier',
                           name: '$'
@@ -3616,7 +3624,7 @@ describe('Next - Private methods', () => {
                         kind: 'set',
                         static: false,
                         computed: false,
-                        decorators: [],
+
                         key: {
                           type: 'PrivateIdentifier',
                           name: '_'
@@ -3676,7 +3684,7 @@ describe('Next - Private methods', () => {
     ],
     [
       `class A { get #foo/*{ declareWith }*/() {} }`,
-      Context.OptionsNext,
+      Context.None,
       {
         body: [
           {
@@ -3684,7 +3692,7 @@ describe('Next - Private methods', () => {
               body: [
                 {
                   computed: false,
-                  decorators: [],
+
                   key: {
                     name: 'foo',
                     type: 'PrivateIdentifier'
@@ -3707,7 +3715,7 @@ describe('Next - Private methods', () => {
               ],
               type: 'ClassBody'
             },
-            decorators: [],
+
             id: {
               name: 'A',
               type: 'Identifier'
@@ -3737,7 +3745,7 @@ describe('Next - Private methods', () => {
         }
 
       }`,
-      Context.OptionsNext,
+      Context.None,
       {
         type: 'Program',
         sourceType: 'script',
@@ -3752,7 +3760,7 @@ describe('Next - Private methods', () => {
                   type: 'ClassExpression',
                   id: null,
                   superClass: null,
-                  decorators: [],
+
                   body: {
                     type: 'ClassBody',
                     body: [
@@ -3765,7 +3773,7 @@ describe('Next - Private methods', () => {
                           type: 'Identifier',
                           name: 'm'
                         },
-                        decorators: [],
+
                         value: {
                           type: 'FunctionExpression',
                           params: [],
@@ -3794,8 +3802,7 @@ describe('Next - Private methods', () => {
                         },
                         value: null,
                         static: false,
-                        computed: false,
-                        decorators: []
+                        computed: false
                       },
                       {
                         type: 'PropertyDefinition',
@@ -3805,15 +3812,14 @@ describe('Next - Private methods', () => {
                         },
                         value: null,
                         static: false,
-                        computed: false,
-                        decorators: []
+                        computed: false
                       },
                       {
                         type: 'MethodDefinition',
                         kind: 'get',
                         static: false,
                         computed: false,
-                        decorators: [],
+
                         key: {
                           type: 'PrivateIdentifier',
                           name: '_'
@@ -3854,7 +3860,7 @@ describe('Next - Private methods', () => {
                           type: 'Identifier',
                           name: '$'
                         },
-                        decorators: [],
+
                         value: {
                           type: 'FunctionExpression',
                           params: [
@@ -3918,7 +3924,7 @@ describe('Next - Private methods', () => {
                           type: 'Identifier',
                           name: '_'
                         },
-                        decorators: [],
+
                         value: {
                           type: 'FunctionExpression',
                           params: [
@@ -3995,7 +4001,7 @@ describe('Next - Private methods', () => {
           return makeEvil(x);
         }
       }`,
-      Context.OptionsNext,
+      Context.None,
       {
         type: 'Program',
         sourceType: 'script',
@@ -4007,7 +4013,7 @@ describe('Next - Private methods', () => {
               name: 'Hotel'
             },
             superClass: null,
-            decorators: [],
+
             body: {
               type: 'ClassBody',
               body: [
@@ -4016,7 +4022,7 @@ describe('Next - Private methods', () => {
                   kind: 'get',
                   static: false,
                   computed: false,
-                  decorators: [],
+
                   key: {
                     type: 'PrivateIdentifier',
                     name: 'evil'
@@ -4050,7 +4056,7 @@ describe('Next - Private methods', () => {
                   kind: 'set',
                   static: false,
                   computed: false,
-                  decorators: [],
+
                   key: {
                     type: 'PrivateIdentifier',
                     name: 'evil'
@@ -4113,7 +4119,7 @@ describe('Next - Private methods', () => {
           return this.#y(43);
         }
       } `,
-      Context.OptionsNext,
+      Context.None,
       {
         body: [
           {
@@ -4128,7 +4134,7 @@ describe('Next - Private methods', () => {
                     body: [
                       {
                         computed: false,
-                        decorators: [],
+
                         key: {
                           name: 'x',
                           type: 'PrivateIdentifier'
@@ -4171,7 +4177,7 @@ describe('Next - Private methods', () => {
                       },
                       {
                         computed: false,
-                        decorators: [],
+
                         key: {
                           name: 'y',
                           type: 'PrivateIdentifier'
@@ -4214,7 +4220,7 @@ describe('Next - Private methods', () => {
                       },
                       {
                         computed: false,
-                        decorators: [],
+
                         key: {
                           name: 'x',
                           type: 'Identifier'
@@ -4260,7 +4266,7 @@ describe('Next - Private methods', () => {
                       },
                       {
                         computed: false,
-                        decorators: [],
+
                         key: {
                           name: 'y',
                           type: 'Identifier'
@@ -4307,7 +4313,7 @@ describe('Next - Private methods', () => {
                     ],
                     type: 'ClassBody'
                   },
-                  decorators: [],
+
                   id: null,
                   superClass: null,
                   type: 'ClassExpression'
@@ -4325,7 +4331,7 @@ describe('Next - Private methods', () => {
     ],
     [
       `class A { static set #foo/*{ declareWith }*/(param) {} }`,
-      Context.OptionsNext,
+      Context.None,
       {
         body: [
           {
@@ -4333,7 +4339,6 @@ describe('Next - Private methods', () => {
               body: [
                 {
                   computed: false,
-                  decorators: [],
                   key: {
                     name: 'foo',
                     type: 'PrivateIdentifier'
@@ -4361,7 +4366,7 @@ describe('Next - Private methods', () => {
               ],
               type: 'ClassBody'
             },
-            decorators: [],
+
             id: {
               name: 'A',
               type: 'Identifier'
@@ -4372,6 +4377,213 @@ describe('Next - Private methods', () => {
         ],
         sourceType: 'script',
         type: 'Program'
+      }
+    ],
+    [
+      `class A { #\\u{61}\nstatic #\\u0062() {} }`,
+      Context.None,
+      {
+        type: 'Program',
+        sourceType: 'script',
+        body: [
+          {
+            type: 'ClassDeclaration',
+            id: {
+              type: 'Identifier',
+              name: 'A'
+            },
+            superClass: null,
+            body: {
+              type: 'ClassBody',
+              body: [
+                {
+                  type: 'PropertyDefinition',
+                  key: {
+                    type: 'PrivateIdentifier',
+                    name: 'a'
+                  },
+                  value: null,
+                  static: false,
+                  computed: false
+                },
+                {
+                  type: 'MethodDefinition',
+                  kind: 'method',
+                  static: true,
+                  computed: false,
+                  key: {
+                    type: 'PrivateIdentifier',
+                    name: 'b'
+                  },
+                  value: {
+                    type: 'FunctionExpression',
+                    params: [],
+                    body: {
+                      type: 'BlockStatement',
+                      body: []
+                    },
+                    async: false,
+                    generator: false,
+                    id: null
+                  }
+                }
+              ]
+            }
+          }
+        ]
+      }
+    ],
+    [
+      'class a {#𐌭人}',
+      Context.OptionsRanges,
+      {
+        type: 'Program',
+        sourceType: 'script',
+        body: [
+          {
+            type: 'ClassDeclaration',
+            id: {
+              type: 'Identifier',
+              name: 'a',
+              start: 6,
+              end: 7,
+              range: [6, 7]
+            },
+            superClass: null,
+            body: {
+              type: 'ClassBody',
+              body: [
+                {
+                  type: 'PropertyDefinition',
+                  key: {
+                    type: 'PrivateIdentifier',
+                    name: '𐌭人',
+                    start: 9,
+                    end: 13,
+                    range: [9, 13]
+                  },
+                  value: null,
+                  static: false,
+                  computed: false,
+                  start: 9,
+                  end: 13,
+                  range: [9, 13]
+                }
+              ],
+              start: 8,
+              end: 14,
+              range: [8, 14]
+            },
+            start: 0,
+            end: 14,
+            range: [0, 14]
+          }
+        ],
+        start: 0,
+        end: 14,
+        range: [0, 14]
+      }
+    ],
+    [
+      'class a {#人}',
+      Context.OptionsRanges,
+      {
+        type: 'Program',
+        sourceType: 'script',
+        body: [
+          {
+            type: 'ClassDeclaration',
+            id: {
+              type: 'Identifier',
+              name: 'a',
+              start: 6,
+              end: 7,
+              range: [6, 7]
+            },
+            superClass: null,
+            body: {
+              type: 'ClassBody',
+              body: [
+                {
+                  type: 'PropertyDefinition',
+                  key: {
+                    type: 'PrivateIdentifier',
+                    name: '人',
+                    start: 9,
+                    end: 11,
+                    range: [9, 11]
+                  },
+                  value: null,
+                  static: false,
+                  computed: false,
+                  start: 9,
+                  end: 11,
+                  range: [9, 11]
+                }
+              ],
+              start: 8,
+              end: 12,
+              range: [8, 12]
+            },
+            start: 0,
+            end: 12,
+            range: [0, 12]
+          }
+        ],
+        start: 0,
+        end: 12,
+        range: [0, 12]
+      }
+    ],
+    [
+      'class a {#𐌭}',
+      Context.OptionsRanges,
+      {
+        type: 'Program',
+        sourceType: 'script',
+        body: [
+          {
+            type: 'ClassDeclaration',
+            id: {
+              type: 'Identifier',
+              name: 'a',
+              start: 6,
+              end: 7,
+              range: [6, 7]
+            },
+            superClass: null,
+            body: {
+              type: 'ClassBody',
+              body: [
+                {
+                  type: 'PropertyDefinition',
+                  key: {
+                    type: 'PrivateIdentifier',
+                    name: '𐌭',
+                    start: 9,
+                    end: 12,
+                    range: [9, 12]
+                  },
+                  value: null,
+                  static: false,
+                  computed: false,
+                  start: 9,
+                  end: 12,
+                  range: [9, 12]
+                }
+              ],
+              start: 8,
+              end: 13,
+              range: [8, 13]
+            },
+            start: 0,
+            end: 13,
+            range: [0, 13]
+          }
+        ],
+        start: 0,
+        end: 13,
+        range: [0, 13]
       }
     ]
   ]);
