@@ -742,6 +742,7 @@ describe('Statements - For of', () => {
                             },
                             right: {
                               type: 'Literal',
+                              // eslint-disable-next-line no-control-regex
                               value: /([--])|[--]|=+|[-$-+-­-]/y,
                               regex: {
                                 pattern: '([--])|[--]|=+|[-\u001c$-+-­-]',
@@ -944,7 +945,8 @@ describe('Statements - For of', () => {
                                 expression: {
                                   type: 'Literal',
                                   value: '¡\u0015\u0001'
-                                }
+                                },
+                                directive: '¡\u0015\u0001'
                               }
                             ]
                           },
